Move NonBlockingGRPCServer to a separate package
It could be used by operator.
